wayland/ipc.c: use single entry point for IPC
