+++ meson.build
@@ -101,6 +101,7 @@ libsrc = [
     'src/config/switcher.c',
     'src/config/toplevel.c',
     'src/wayland/foreign-toplevel.c',
+    'src/wayland/ipc.c',
     'src/wayland/probe.c',
     'src/wayland/xdg-output.c',
     wayland_targets ]
